変更

+++ Planet/Assets/Scripts/Oura/ExplosionDamage.cs
@@ -7,7 +7,7 @@ public class ExplosionDamage : MonoBehaviour
     private int playerDamage = 30;
 
     [SerializeField]
-    private float planetDamage = 20f;
+    private int planetDamage = 20;
 
     // 判定時間
     [SerializeField]
